refactor(frontend): converge loading/empty states onto shared components (R36)
Port CardSkeleton/GridSkeleton/FadeIn/StaggeredFadeIn into shared
ClinicalLoadingState; repoint ChartReviewTabOptimized empty states onto
ClinicalEmptyState; convert CDSHooksTab spinner loading branches to the
shared ResourceCard skeleton; drop dead Enhanced imports from
EnhancedOrdersTab; delete EnhancedLoadingStates/EnhancedEmptyStates.
