+++ frontend/src/components/clinical/workspace/tabs/EnhancedOrdersTab.js
@@ -70,22 +70,12 @@ import { getStatusColor, getSeverityColor } from '../../../../themes/clinicalThe
 import websocketService from '../../../../services/websocket';
 
 // Enhanced UX components
-import { 
-  CardSkeleton, 
-  GridSkeleton, 
-  FadeInContainer, 
-  StaggeredFadeIn,
-  LoadingOverlay 
-} from './components/EnhancedLoadingStates';
-import { 
-  InteractiveIconButton, 
-  InteractiveButton, 
-  InteractiveChip,
-  RichTooltip 
-} from './components/EnhancedInteractions';
 import {
-  HelpfulEmptyState
-} from './components/EnhancedEmptyStates';
+  InteractiveIconButton,
+  InteractiveButton,
+  InteractiveChip,
+  RichTooltip
+} from './components/EnhancedInteractions';
 
 // Existing dialogs
 // CPOEDialog is retained for single-order EDIT mode only (Phase 4.4
